be/pyproject.toml
[tool.poetry]
package-mode = false
name = "giggrove"
version = "0.1.0"
description = "Automate Job Tracking with CrewAI"

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
crewai = {extras = ["tools"], version = "0.201.1"}  
pydantic = "^2.6.3"
load-dotenv = "^0.1.0"
crewai-tools = "^0.0.15"
flask = "^3.0.2"
flask-cors = "^4.0.0"
setuptools = "<81"  

[tool.pyright]
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
